test: adopt shared assert/source-contract helpers and real-asset gating across the suite
Coordinated test-hardening sweep: adopt shared VectorAssert/SourceContract helpers, route hardcoded BSA paths through SampleBsaLocator, gate real-asset/GPU/shader tests via BucketBTestGuard/GpuTestGuard/ShaderCompileTestGuard, add [Collection(SequentialIntegrationGroup)] to real-asset classes, and migrate temp-file fixtures to in-memory MemoryMappedFile. Includes the new shared helper sources. No production behavior change.

+++ tests/BethesdaMultitool.Tests/App/MapMarkerAutomationLayoutTests.cs
@@ -16,8 +16,8 @@ public sealed class MapMarkerAutomationLayoutTests
             200f, 200f, metrics, 1f, hasIcon: true);
 
         Assert.NotNull(bounds);
-        Assert.Equal(24f, (float)bounds.Value.Width, 3);
-        Assert.Equal(24f, (float)bounds.Value.Height, 3);
+        Assert.Equal(24f, bounds.Value.Width, 3);
+        Assert.Equal(24f, bounds.Value.Height, 3);
     }
 
     [Fact]
@@ -30,8 +30,8 @@ public sealed class MapMarkerAutomationLayoutTests
             200f, 200f, metrics, 1f, hasIcon: true);
 
         Assert.NotNull(bounds);
-        Assert.Equal(40f, (float)bounds.Value.Width, 3);
-        Assert.Equal(40f, (float)bounds.Value.Height, 3);
+        Assert.Equal(40f, bounds.Value.Width, 3);
+        Assert.Equal(40f, bounds.Value.Height, 3);
     }
 
     [Fact]
@@ -44,8 +44,8 @@ public sealed class MapMarkerAutomationLayoutTests
             200f, 200f, metrics, 2f, hasIcon: true);
 
         Assert.NotNull(bounds);
-        Assert.Equal(48f, (float)bounds.Value.Width, 3);
-        Assert.Equal(40f, (float)bounds.Value.Height, 3);
+        Assert.Equal(48f, bounds.Value.Width, 3);
+        Assert.Equal(40f, bounds.Value.Height, 3);
     }
 
     [Fact]
